web/Uzmeklesana/requirements.txt
click
colorama
Flask
itsdangerous
Jinja2
MarkupSafe
Werkzeug
gunicorn
Flask-MySQL
pymysql
mysql-connector-python
cryptography
validators
requests
